Fix many more implicit loss of precision warnings in Monet.

+++ Applications/Monet/MDataEntryController.h
@@ -38,7 +38,7 @@
 - (IBAction)removeSymbol:(id)sender;
 
 // NSTableView data source
-- (int)numberOfRowsInTableView:(NSTableView *)tableView;
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)tableView;
 - (id)tableView:(NSTableView *)tableView objectValueForTableColumn:(NSTableColumn *)tableColumn row:(NSInteger)row;
 - (void)tableView:(NSTableView *)tableView setObjectValue:(id)object forTableColumn:(NSTableColumn *)tableColumn row:(NSInteger)row;
